Split configuration tests
